pyproject.toml
[project]
name = "moodmate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Amdubaloch", email = "amdubaloch2005@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.14",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
]

[project.scripts]
moodmate = "moodmate:main"
mood = "moodmate.moodmate:moodmate"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
